+++ Public/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### Release 0.1.0 Version
+
+-   Added save time randomizer. By default is on. Randomizes backup cooldown time from 24 - 168 hours. Can be modified through configs.
+-   Added mods communicator 4 events: `SaveBackupBefore`, `SaveBackupAfter`, `DeathRollBefore`, `DeathRollAfter`.
+-   Removed marker files creation and their logic. We already storing, which character is restored, inside metadata.
+
 ### Release 0.0.2 Version
 
 -   Fixed description. Now it show correct shield badges.